src/pyproject.toml
[tool.poetry]
name = "forecast"
version = "0.1.0"
description = "Challenge technical exam for coderio."
authors = ["redraw <agustinbv@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.4"
requests = "^2.23.0"
djangorestframework = "^3.11.0"
markdown = "^3.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
